Extract configuration building to pgmgr/config.go.
